+++ x/staking/keeper/delegation.go
@@ -695,7 +695,10 @@ func (k Keeper) SetRedelegationQueueTimeSlice(ctx context.Context, timestamp tim
 // InsertRedelegationQueue insert an redelegation delegation to the appropriate
 // timeslice in the redelegation queue.
 func (k Keeper) InsertRedelegationQueue(ctx context.Context, red types.Redelegation, completionTime time.Time) error {
-	timeSlice, _ := k.GetRedelegationQueueTimeSlice(ctx, completionTime)
+	timeSlice, err := k.GetRedelegationQueueTimeSlice(ctx, completionTime)
+	if err != nil {
+		return err
+	}
 	dvvTriplet := types.DVVTriplet{
 		DelegatorAddress:    red.DelegatorAddress,
 		ValidatorSrcAddress: red.ValidatorSrcAddress,